+++ src/runtime/RuntimeHelperInterface.php
@@ -49,11 +49,11 @@ interface RuntimeHelperInterface
     /**
      * @param string $element
      * @param array $attributes
-     * @param array|null $content
+     * @param string|null $content
      * @return string
      * @since 0.1.0
      */
-    public static function createElement(string $element, array $attributes = [], ?array $content = null): string;
+    public static function createElement(string $element, array $attributes = [], ?string $content = null): string;
 
     /**
      * @param string $name