+++ tests/unit/compile/cases/components/03.runtime-error.t
@@ -6,7 +6,7 @@
     </Step>
 </Step>
 ---- CODE ----
-($runtime->createComponent('Step',[],function($runtime){return [($runtime->createComponent('Step',[],function($runtime){return [($runtime->createComponent('TestComponent',[],function($runtime){return [($runtime->createComponent('Failure',['message'=>'Lorem ipsum dolor']))];}))];}))];}))
+($runtime->createComponent('Step',[],function($runtime){return ($runtime->createComponent('Step',[],function($runtime){return ($runtime->createComponent('TestComponent',[],function($runtime){return ($runtime->createComponent('Failure',['message'=>'Lorem ipsum dolor']));}));}));}))
 ---- THROW ----
 An error occurred while executing template
 -- prev --