+++ tests/unit/compile/cases/html/12.component-block-empty-props.t
@@ -1,5 +1,5 @@
 Lorem <TestComponent foo=42 bar={lol}></TestComponent> ipsum
 ---- CODE ----
-('Lorem '.($runtime->createComponent('TestComponent',['foo'=>'42','bar'=>($runtime->param('lol'))],function($runtime){return [];})).' ipsum')
+('Lorem '.($runtime->createComponent('TestComponent',['foo'=>'42','bar'=>($runtime->param('lol'))],function($runtime){return '';})).' ipsum')
 ---- RESULT ----
 Lorem <!-- Test Component: foo="42" bar="[value of &lol]" --><!-- /Test Component --> ipsum