+++ tests/unit/compile/cases/tags/expr/12.attr-inject-qq.t
@@ -1,5 +1,5 @@
 <div title={ "Lorem${ foo }ipsum&amp;" }></div>
 ---- CODE ----
-($runtime::createElement('div',['title'=>('Lorem'.($runtime::toString(($runtime->param('foo')))).'ipsum&amp;')],[]))
+($runtime::createElement('div',['title'=>('Lorem'.($runtime::toString(($runtime->param('foo')))).'ipsum&amp;')],''))
 ---- RESULT ----
 <div title="Lorem[value of &amp;foo]ipsum&amp;amp;"></div>